changed to float, issues with loss value

+++ host/include/loss.h
@@ -3,13 +3,13 @@
 
 #include "matrix.h"
 
-double multiclass_softmax(matrix_t* pred, matrix_t* label);
+float multiclass_softmax(matrix_t* pred, matrix_t* label);
 
-void softmax_single(double* arr, int size);
+void softmax_single(float* arr, int size);
 
 void softmax(matrix_t* m);
 
-double mean_cross_entropy_softmax(matrix_t* logits, matrix_t* labels);
+float mean_cross_entropy_softmax(matrix_t* logits, matrix_t* labels);
 
 matrix_t* d_mean_cross_entropy_softmax(matrix_t* logits, matrix_t* labels);
