+++ host/src/matrix.c
@@ -12,13 +12,13 @@ matrix_t* create_matrix(int rows, int cols) {
     m->rows = rows;
     m->cols = cols;
 
-    m->vals = calloc(m->rows, sizeof(double*));
+    m->vals = calloc(m->rows, sizeof(float*));
     if (!m->vals) {
         destroy_matrix(m);
         assert(false);
     }
     for (int i=0; i<m->rows; ++i) {
-        double* row = calloc(m->cols, sizeof(double));
+        float* row = calloc(m->cols, sizeof(float));
         if (!row) {
             destroy_matrix(m);
             assert(false);
@@ -44,7 +44,7 @@ matrix_t* create_matrix_identity(int dim) {
 }
 
 // create matrix initialized with values in range randomly
-matrix_t* create_matrix_random(int rows, int cols, double start, double end) {
+matrix_t* create_matrix_random(int rows, int cols, float start, float end) {
     assert(end > start);
 
     matrix_t* m = create_matrix(rows, cols);
@@ -54,7 +54,7 @@ matrix_t* create_matrix_random(int rows, int cols, double start, double end) {
 
     for (int i=0; i<m->rows; ++i) {
         for (int j=0; j<m->cols; ++j) {
-            double val = ((double)rand())/((double)RAND_MAX) * (end - start) + start;
+            float val = ((float)rand())/((float)RAND_MAX) * (end - start) + start;
             m->vals[i][j] = val;
         }
     }
@@ -255,7 +255,7 @@ matrix_t* transpose_matrix(matrix_t* m) {
 // given a function that takes in a single value and 
 // returns a single value, apply that function to every
 // element of the matrix
-void apply_matrix(double (*apply)(double), matrix_t* m){
+void apply_matrix(float (*apply)(float), matrix_t* m){
     if (!m) {
         assert(false);
     }
