+++ host/src/network.c
@@ -19,7 +19,7 @@ void init_network(void) {
     int layers[] = {784, 64, 32, 10};
     int loaded_start = 0;
     int n_loaded = 3;
-    double learning_rate = 0.03;
+    float learning_rate = 0.03;
     int batch_size = 50;
     optimizer_type_t optimizer = GD;
 
@@ -155,7 +155,7 @@ void destroy_layer(layer_t* layer) {
 
 // input and labels are in row-major order, meaning 
 // each row corresponds to a particular training example. 
-double forward(matrix_t* features, matrix_t* labels) {
+float forward(matrix_t* features, matrix_t* labels) {
     //DMSG("starting forward propagation\n");
     assert(nn != NULL && features != NULL);
 
@@ -303,7 +303,7 @@ void train(int epochs) {
     for (int epoch=0; epoch < epochs; ++epoch) {
         
         
-        double sum_loss = 0.0;
+        float sum_loss = 0.0;
         int batch_size = nn->batch_size;
         int b = 0;
         for (b=0; b < ((data_loader->N - 1) / batch_size + 1); ++b) { // iterate through batches
@@ -317,7 +317,7 @@ void train(int epochs) {
                                                                         data_loader->labels->rows),
                                                     0, data_loader->labels->cols);
 
-            double loss = forward(batch_features, batch_labels);
+            float loss = forward(batch_features, batch_labels);
             // if (isnan(loss) || isinf(loss)) {
             //     //DMSG("abnormal loss in epoch %d, batch %d\n", epoch, b);
             //     continue;
@@ -336,7 +336,7 @@ void train(int epochs) {
         }
         sum_loss /= (b + 1);
         matrix_t* y_hat = predict(data_loader->features);
-        double acc = accuracy(y_hat, data_loader->labels);
+        float acc = accuracy(y_hat, data_loader->labels);
         destroy_matrix(y_hat);
         
         printf("========= Epoch: %d, Loss: %f, Accuracy: %f ==========", epoch, sum_loss, acc);
@@ -380,13 +380,13 @@ matrix_t* predict(matrix_t* features) {
 
 
 // given predictions and labels, calculate accuracy
-double accuracy(matrix_t* y_hat, matrix_t* labels) {
+float accuracy(matrix_t* y_hat, matrix_t* labels) {
     assert(y_hat->rows == labels->rows && y_hat->cols == labels->cols);
 
-    double correct = 0;
+    float correct = 0;
     for (int i=0; i < y_hat->rows; ++i) {
         int max_index = 0;
-        double max_value = 0.0;
+        float max_value = 0.0;
         for (int j=0; j < y_hat->cols; ++j) {
             if (y_hat->vals[i][j] > max_value) {
                 max_value = y_hat->vals[i][j];
@@ -396,7 +396,7 @@ double accuracy(matrix_t* y_hat, matrix_t* labels) {
         correct += labels->vals[i][max_index];
     }
 
-    return correct / ((double) (y_hat->rows));
+    return correct / ((float) (y_hat->rows));
 }
 
 // sawp the current layers in nn with the new layers
