+++ ta/activations.c
@@ -1,19 +1,19 @@
 #include <activations.h>
 #include <math_TA.h>
 
-double sigmoid(double x) {
+float sigmoid(float x) {
     return 1 / (1 + ta_exp(-x));
 }
 
-double d_sigmoid(double x) {
+float d_sigmoid(float x) {
     return sigmoid(x) * (1 - sigmoid(x));
 }
 
-double relu(double x) {
+float relu(float x) {
     return MAX(0.0, x);
 }
 
-double d_relu(double x) {
+float d_relu(float x) {
     if (x < 0) {
         return 0.0;
     }
